Move CI tests to appveyor
Unfortunately, github does not disable SIP on their build machines,
which makes it impossible for us to run our unit tests without signing
them. Further complicating things, we can't limit the signature to a
development signature, because we do not know the ID of the macOS
machines our tests run on.

Appveyor's machines have SIP disable, so let's rely on them.